paginate channel index
