+++ src/webserver/messages.py
@@ -1,9 +1,6 @@
-from django.conf import settings
-from pymongo import MongoClient
-from urllib.parse import quote_plus
+from webserver.db import *
 from math import ceil
 
-DEFAULT_LIMIT = 50
 DEFAULT_FIELDS = [
     'channel',
     'username',
@@ -16,9 +13,6 @@ DEFAULT_FIELDS = [
     'emotes_seventv'
 ]
 DEFAULT_SORT = list({'timestamp': -1}.items())
-MESSAGE_COLLECTION = 'messages'
-CHANNEL_COLLECTION = 'channels'
-TIMESTAMP_FORMAT = '%H:%M:%S %m/%d/%y'
     
 def get_channel_messages(channel:str, filter={}, sort=DEFAULT_SORT, fields=DEFAULT_FIELDS, limit=DEFAULT_LIMIT, page=0):
     container = get_db()[MESSAGE_COLLECTION]
@@ -58,19 +52,6 @@ def get_page_count(channel=None, username=None, filter={}, limit=DEFAULT_LIMIT):
     total = container.count_documents(filter=filter)
     return ceil(total / limit)
 
-def get_channels():
-    #return get_db()[MESSAGE_COLLECTION].distinct('channel')
-    project = {
-        '_id': 0,
-        'channel': 1,
-        'message_count': 1
-    }
-    sort = list({'channel': 1}.items())
-    return [x for x in get_db()[CHANNEL_COLLECTION].find(
-        projection=project,
-        sort=sort
-    )]
-
 def get_user_color(username):
     username = username.lower()
     collection = get_db()[MESSAGE_COLLECTION]
@@ -107,50 +88,6 @@ def get_user_messages(username, channels=None, filter={}, sort=DEFAULT_SORT, fie
         limit=limit
     ).limit(limit)
     return parse_messages(cursor)
-
-def get_project(fields):
-    project = {}
-    project['_id'] = 0
-    for field in fields:
-        project[field] = 1
-    if 'username' in fields:
-        project['color'] = 1
-    return project
-
-def get_db():
-    return get_client()[settings.MONGO_DBNAME]
-
-def get_client():
-    return MongoClient(get_connection_string(
-        dbhosts=settings.MONGO_HOSTS,
-        dbusername=settings.MONGO_USER,
-        dbpassword=settings.MONGO_PASS,
-        defaultauthdb=settings.MONGO_AUTHDB,
-        dboptions=settings.MONGO_OPTIONS
-    ))
-
-def get_connection_string(dbhosts, dbusername, dbpassword, defaultauthdb, dboptions):
-        out_string = "mongodb://"
-        if dbusername:
-            out_string += f'{quote_plus(dbusername)}:{quote_plus(dbpassword)}@'
-        for i in range(len(dbhosts)):
-            if i > 0:
-                out_string += ','
-            out_string += f'{dbhosts[i][0]}'
-            if dbhosts[i][1]:
-                out_string += f':{dbhosts[i][1]}'
-        out_string += '/'
-        if defaultauthdb:
-            out_string += defaultauthdb
-        if dboptions:
-            out_string += '?'
-            option_count = 0
-            for option in dboptions:
-                if option_count > 0:
-                    out_string += '&'
-                out_string += f'{option}={dboptions[option]}'
-                option_count+=1
-        return out_string
 
 def parse_messages(cursor):
     messages = []
